+++ packages/paraglide/paraglide-astro/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/paraglide-astro
 
+## 0.1.5
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.45
+
+## 0.1.4
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.44
+
 ## 0.1.3
 
 ### Patch Changes